+++ tests/message/message_conformance.pb.go
@@ -25,7 +25,7 @@ const (
 // of the legacy proto package is being used.
 const _ = proto.ProtoPackageIsVersion4
 
-type BasicMessage struct {
+type Basic struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
@@ -34,8 +34,8 @@ type BasicMessage struct {
 	Name string `protobuf:"bytes,2,opt,name=name,proto3" json:"name,omitempty"`
 }
 
-func (x *BasicMessage) Reset() {
-	*x = BasicMessage{}
+func (x *Basic) Reset() {
+	*x = Basic{}
 	if protoimpl.UnsafeEnabled {
 		mi := &file_tests_message_message_conformance_proto_msgTypes[0]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -43,13 +43,13 @@ func (x *BasicMessage) Reset() {
 	}
 }
 
-func (x *BasicMessage) String() string {
+func (x *Basic) String() string {
 	return protoimpl.X.MessageStringOf(x)
 }
 
-func (*BasicMessage) ProtoMessage() {}
+func (*Basic) ProtoMessage() {}
 
-func (x *BasicMessage) ProtoReflect() protoreflect.Message {
+func (x *Basic) ProtoReflect() protoreflect.Message {
 	mi := &file_tests_message_message_conformance_proto_msgTypes[0]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -61,38 +61,38 @@ func (x *BasicMessage) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use BasicMessage.ProtoReflect.Descriptor instead.
-func (*BasicMessage) Descriptor() ([]byte, []int) {
+// Deprecated: Use Basic.ProtoReflect.Descriptor instead.
+func (*Basic) Descriptor() ([]byte, []int) {
 	return file_tests_message_message_conformance_proto_rawDescGZIP(), []int{0}
 }
 
-func (x *BasicMessage) GetId() int32 {
+func (x *Basic) GetId() int32 {
 	if x != nil {
 		return x.Id
 	}
 	return 0
 }
 
-func (x *BasicMessage) GetName() string {
+func (x *Basic) GetName() string {
 	if x != nil {
 		return x.Name
 	}
 	return ""
 }
 
-type OneofMessage struct {
+type Union struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
 	// Types that are assignable to Contents:
-	//	*OneofMessage_Id
-	//	*OneofMessage_Name
-	Contents isOneofMessage_Contents `protobuf_oneof:"contents"`
+	//	*Union_Id
+	//	*Union_Name
+	Contents isUnion_Contents `protobuf_oneof:"contents"`
 }
 
-func (x *OneofMessage) Reset() {
-	*x = OneofMessage{}
+func (x *Union) Reset() {
+	*x = Union{}
 	if protoimpl.UnsafeEnabled {
 		mi := &file_tests_message_message_conformance_proto_msgTypes[1]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -100,13 +100,13 @@ func (x *OneofMessage) Reset() {
 	}
 }
 
-func (x *OneofMessage) String() string {
+func (x *Union) String() string {
 	return protoimpl.X.MessageStringOf(x)
 }
 
-func (*OneofMessage) ProtoMessage() {}
+func (*Union) ProtoMessage() {}
 
-func (x *OneofMessage) ProtoReflect() protoreflect.Message {
+func (x *Union) ProtoReflect() protoreflect.Message {
 	mi := &file_tests_message_message_conformance_proto_msgTypes[1]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -118,58 +118,59 @@ func (x *OneofMessage) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use OneofMessage.ProtoReflect.Descriptor instead.
-func (*OneofMessage) Descriptor() ([]byte, []int) {
+// Deprecated: Use Union.ProtoReflect.Descriptor instead.
+func (*Union) Descriptor() ([]byte, []int) {
 	return file_tests_message_message_conformance_proto_rawDescGZIP(), []int{1}
 }
 
-func (m *OneofMessage) GetContents() isOneofMessage_Contents {
+func (m *Union) GetContents() isUnion_Contents {
 	if m != nil {
 		return m.Contents
 	}
 	return nil
 }
 
-func (x *OneofMessage) GetId() int32 {
-	if x, ok := x.GetContents().(*OneofMessage_Id); ok {
+func (x *Union) GetId() int32 {
+	if x, ok := x.GetContents().(*Union_Id); ok {
 		return x.Id
 	}
 	return 0
 }
 
-func (x *OneofMessage) GetName() string {
-	if x, ok := x.GetContents().(*OneofMessage_Name); ok {
+func (x *Union) GetName() string {
+	if x, ok := x.GetContents().(*Union_Name); ok {
 		return x.Name
 	}
 	return ""
 }
 
-type isOneofMessage_Contents interface {
-	isOneofMessage_Contents()
+type isUnion_Contents interface {
+	isUnion_Contents()
 }
 
-type OneofMessage_Id struct {
+type Union_Id struct {
 	Id int32 `protobuf:"varint,1,opt,name=id,proto3,oneof"`
 }
 
-type OneofMessage_Name struct {
+type Union_Name struct {
 	Name string `protobuf:"bytes,2,opt,name=name,proto3,oneof"`
 }
 
-func (*OneofMessage_Id) isOneofMessage_Contents() {}
+func (*Union_Id) isUnion_Contents() {}
 
-func (*OneofMessage_Name) isOneofMessage_Contents() {}
+func (*Union_Name) isUnion_Contents() {}
 
-type OuterMessage struct {
+type Outer struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Inner *OuterMessage_InnerMessage `protobuf:"bytes,1,opt,name=inner,proto3" json:"inner,omitempty"`
+	Middle *Outer_Middle       `protobuf:"bytes,1,opt,name=middle,proto3" json:"middle,omitempty"`
+	Inner  *Outer_Middle_Inner `protobuf:"bytes,2,opt,name=inner,proto3" json:"inner,omitempty"`
 }
 
-func (x *OuterMessage) Reset() {
-	*x = OuterMessage{}
+func (x *Outer) Reset() {
+	*x = Outer{}
 	if protoimpl.UnsafeEnabled {
 		mi := &file_tests_message_message_conformance_proto_msgTypes[2]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -177,13 +178,13 @@ func (x *OuterMessage) Reset() {
 	}
 }
 
-func (x *OuterMessage) String() string {
+func (x *Outer) String() string {
 	return protoimpl.X.MessageStringOf(x)
 }
 
-func (*OuterMessage) ProtoMessage() {}
+func (*Outer) ProtoMessage() {}
 
-func (x *OuterMessage) ProtoReflect() protoreflect.Message {
+func (x *Outer) ProtoReflect() protoreflect.Message {
 	mi := &file_tests_message_message_conformance_proto_msgTypes[2]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -195,26 +196,35 @@ func (x *OuterMessage) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use OuterMessage.ProtoReflect.Descriptor instead.
-func (*OuterMessage) Descriptor() ([]byte, []int) {
+// Deprecated: Use Outer.ProtoReflect.Descriptor instead.
+func (*Outer) Descriptor() ([]byte, []int) {
 	return file_tests_message_message_conformance_proto_rawDescGZIP(), []int{2}
 }
 
-func (x *OuterMessage) GetInner() *OuterMessage_InnerMessage {
+func (x *Outer) GetMiddle() *Outer_Middle {
+	if x != nil {
+		return x.Middle
+	}
+	return nil
+}
+
+func (x *Outer) GetInner() *Outer_Middle_Inner {
 	if x != nil {
 		return x.Inner
 	}
 	return nil
 }
 
-type OuterMessage_InnerMessage struct {
+type Outer_Middle struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
+
+	Inner *Outer_Middle_Inner `protobuf:"bytes,1,opt,name=inner,proto3" json:"inner,omitempty"`
 }
 
-func (x *OuterMessage_InnerMessage) Reset() {
-	*x = OuterMessage_InnerMessage{}
+func (x *Outer_Middle) Reset() {
+	*x = Outer_Middle{}
 	if protoimpl.UnsafeEnabled {
 		mi := &file_tests_message_message_conformance_proto_msgTypes[3]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -222,13 +232,13 @@ func (x *OuterMessage_InnerMessage) Reset() {
 	}
 }
 
-func (x *OuterMessage_InnerMessage) String() string {
+func (x *Outer_Middle) String() string {
 	return protoimpl.X.MessageStringOf(x)
 }
 
-func (*OuterMessage_InnerMessage) ProtoMessage() {}
+func (*Outer_Middle) ProtoMessage() {}
 
-func (x *OuterMessage_InnerMessage) ProtoReflect() protoreflect.Message {
+func (x *Outer_Middle) ProtoReflect() protoreflect.Message {
 	mi := &file_tests_message_message_conformance_proto_msgTypes[3]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -240,9 +250,63 @@ func (x *OuterMessage_InnerMessage) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use OuterMessage_InnerMessage.ProtoReflect.Descriptor instead.
-func (*OuterMessage_InnerMessage) Descriptor() ([]byte, []int) {
+// Deprecated: Use Outer_Middle.ProtoReflect.Descriptor instead.
+func (*Outer_Middle) Descriptor() ([]byte, []int) {
 	return file_tests_message_message_conformance_proto_rawDescGZIP(), []int{2, 0}
+}
+
+func (x *Outer_Middle) GetInner() *Outer_Middle_Inner {
+	if x != nil {
+		return x.Inner
+	}
+	return nil
+}
+
+type Outer_Middle_Inner struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Id int32 `protobuf:"varint,1,opt,name=id,proto3" json:"id,omitempty"`
+}
+
+func (x *Outer_Middle_Inner) Reset() {
+	*x = Outer_Middle_Inner{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_tests_message_message_conformance_proto_msgTypes[4]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *Outer_Middle_Inner) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Outer_Middle_Inner) ProtoMessage() {}
+
+func (x *Outer_Middle_Inner) ProtoReflect() protoreflect.Message {
+	mi := &file_tests_message_message_conformance_proto_msgTypes[4]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Outer_Middle_Inner.ProtoReflect.Descriptor instead.
+func (*Outer_Middle_Inner) Descriptor() ([]byte, []int) {
+	return file_tests_message_message_conformance_proto_rawDescGZIP(), []int{2, 0, 0}
+}
+
+func (x *Outer_Middle_Inner) GetId() int32 {
+	if x != nil {
+		return x.Id
+	}
+	return 0
 }
 
 var File_tests_message_message_conformance_proto protoreflect.FileDescriptor
@@ -251,24 +315,30 @@ var file_tests_message_message_conformance_proto_rawDesc = []byte{
 	0x0a, 0x27, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2f, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2f,
 	0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x5f, 0x63, 0x6f, 0x6e, 0x66, 0x6f, 0x72, 0x6d, 0x61,
 	0x6e, 0x63, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0d, 0x74, 0x65, 0x73, 0x74, 0x73,
-	0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x22, 0x32, 0x0a, 0x0c, 0x42, 0x61, 0x73, 0x69,
-	0x63, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x05, 0x52, 0x02, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x22, 0x42, 0x0a, 0x0c,
-	0x4f, 0x6e, 0x65, 0x6f, 0x66, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x10, 0x0a, 0x02,
-	0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x48, 0x00, 0x52, 0x02, 0x69, 0x64, 0x12, 0x14,
-	0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x04,
-	0x6e, 0x61, 0x6d, 0x65, 0x42, 0x0a, 0x0a, 0x08, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x73,
-	0x22, 0x5e, 0x0a, 0x0c, 0x4f, 0x75, 0x74, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
-	0x12, 0x3e, 0x0a, 0x05, 0x69, 0x6e, 0x6e, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x28, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e,
-	0x4f, 0x75, 0x74, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x49, 0x6e, 0x6e,
-	0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x05, 0x69, 0x6e, 0x6e, 0x65, 0x72,
-	0x1a, 0x0e, 0x0a, 0x0c, 0x49, 0x6e, 0x6e, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
-	0x42, 0x2a, 0x5a, 0x28, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x61,
-	0x6c, 0x74, 0x61, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x70, 0x61, 0x74, 0x63, 0x68, 0x2f, 0x74,
-	0x65, 0x73, 0x74, 0x73, 0x2f, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x62, 0x06, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x33,
+	0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x22, 0x2b, 0x0a, 0x05, 0x42, 0x61, 0x73, 0x69,
+	0x63, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x02, 0x69,
+	0x64, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x04, 0x6e, 0x61, 0x6d, 0x65, 0x22, 0x3b, 0x0a, 0x05, 0x55, 0x6e, 0x69, 0x6f, 0x6e, 0x12, 0x10,
+	0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x48, 0x00, 0x52, 0x02, 0x69, 0x64,
+	0x12, 0x14, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00,
+	0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x42, 0x0a, 0x0a, 0x08, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e,
+	0x74, 0x73, 0x22, 0xd1, 0x01, 0x0a, 0x05, 0x4f, 0x75, 0x74, 0x65, 0x72, 0x12, 0x33, 0x0a, 0x06,
+	0x6d, 0x69, 0x64, 0x64, 0x6c, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x74,
+	0x65, 0x73, 0x74, 0x73, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x4f, 0x75, 0x74,
+	0x65, 0x72, 0x2e, 0x4d, 0x69, 0x64, 0x64, 0x6c, 0x65, 0x52, 0x06, 0x6d, 0x69, 0x64, 0x64, 0x6c,
+	0x65, 0x12, 0x37, 0x0a, 0x05, 0x69, 0x6e, 0x6e, 0x65, 0x72, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b,
+	0x32, 0x21, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x2e, 0x4f, 0x75, 0x74, 0x65, 0x72, 0x2e, 0x4d, 0x69, 0x64, 0x64, 0x6c, 0x65, 0x2e, 0x49, 0x6e,
+	0x6e, 0x65, 0x72, 0x52, 0x05, 0x69, 0x6e, 0x6e, 0x65, 0x72, 0x1a, 0x5a, 0x0a, 0x06, 0x4d, 0x69,
+	0x64, 0x64, 0x6c, 0x65, 0x12, 0x37, 0x0a, 0x05, 0x69, 0x6e, 0x6e, 0x65, 0x72, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x0b, 0x32, 0x21, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2e, 0x6d, 0x65, 0x73, 0x73,
+	0x61, 0x67, 0x65, 0x2e, 0x4f, 0x75, 0x74, 0x65, 0x72, 0x2e, 0x4d, 0x69, 0x64, 0x64, 0x6c, 0x65,
+	0x2e, 0x49, 0x6e, 0x6e, 0x65, 0x72, 0x52, 0x05, 0x69, 0x6e, 0x6e, 0x65, 0x72, 0x1a, 0x17, 0x0a,
+	0x05, 0x49, 0x6e, 0x6e, 0x65, 0x72, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01,
+	0x28, 0x05, 0x52, 0x02, 0x69, 0x64, 0x42, 0x2a, 0x5a, 0x28, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62,
+	0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x61, 0x6c, 0x74, 0x61, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x70,
+	0x61, 0x74, 0x63, 0x68, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2f, 0x6d, 0x65, 0x73, 0x73, 0x61,
+	0x67, 0x65, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -283,20 +353,23 @@ func file_tests_message_message_conformance_proto_rawDescGZIP() []byte {
 	return file_tests_message_message_conformance_proto_rawDescData
 }
 
-var file_tests_message_message_conformance_proto_msgTypes = make([]protoimpl.MessageInfo, 4)
+var file_tests_message_message_conformance_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
 var file_tests_message_message_conformance_proto_goTypes = []interface{}{
-	(*BasicMessage)(nil),              // 0: tests.message.BasicMessage
-	(*OneofMessage)(nil),              // 1: tests.message.OneofMessage
-	(*OuterMessage)(nil),              // 2: tests.message.OuterMessage
-	(*OuterMessage_InnerMessage)(nil), // 3: tests.message.OuterMessage.InnerMessage
+	(*Basic)(nil),              // 0: tests.message.Basic
+	(*Union)(nil),              // 1: tests.message.Union
+	(*Outer)(nil),              // 2: tests.message.Outer
+	(*Outer_Middle)(nil),       // 3: tests.message.Outer.Middle
+	(*Outer_Middle_Inner)(nil), // 4: tests.message.Outer.Middle.Inner
 }
 var file_tests_message_message_conformance_proto_depIdxs = []int32{
-	3, // 0: tests.message.OuterMessage.inner:type_name -> tests.message.OuterMessage.InnerMessage
-	1, // [1:1] is the sub-list for method output_type
-	1, // [1:1] is the sub-list for method input_type
-	1, // [1:1] is the sub-list for extension type_name
-	1, // [1:1] is the sub-list for extension extendee
-	0, // [0:1] is the sub-list for field type_name
+	3, // 0: tests.message.Outer.middle:type_name -> tests.message.Outer.Middle
+	4, // 1: tests.message.Outer.inner:type_name -> tests.message.Outer.Middle.Inner
+	4, // 2: tests.message.Outer.Middle.inner:type_name -> tests.message.Outer.Middle.Inner
+	3, // [3:3] is the sub-list for method output_type
+	3, // [3:3] is the sub-list for method input_type
+	3, // [3:3] is the sub-list for extension type_name
+	3, // [3:3] is the sub-list for extension extendee
+	0, // [0:3] is the sub-list for field type_name
 }
 
 func init() { file_tests_message_message_conformance_proto_init() }
@@ -306,7 +379,7 @@ func file_tests_message_message_conformance_proto_init() {
 	}
 	if !protoimpl.UnsafeEnabled {
 		file_tests_message_message_conformance_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*BasicMessage); i {
+			switch v := v.(*Basic); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -318,7 +391,7 @@ func file_tests_message_message_conformance_proto_init() {
 			}
 		}
 		file_tests_message_message_conformance_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*OneofMessage); i {
+			switch v := v.(*Union); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -330,7 +403,7 @@ func file_tests_message_message_conformance_proto_init() {
 			}
 		}
 		file_tests_message_message_conformance_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*OuterMessage); i {
+			switch v := v.(*Outer); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -342,7 +415,19 @@ func file_tests_message_message_conformance_proto_init() {
 			}
 		}
 		file_tests_message_message_conformance_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*OuterMessage_InnerMessage); i {
+			switch v := v.(*Outer_Middle); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_tests_message_message_conformance_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*Outer_Middle_Inner); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -355,8 +440,8 @@ func file_tests_message_message_conformance_proto_init() {
 		}
 	}
 	file_tests_message_message_conformance_proto_msgTypes[1].OneofWrappers = []interface{}{
-		(*OneofMessage_Id)(nil),
-		(*OneofMessage_Name)(nil),
+		(*Union_Id)(nil),
+		(*Union_Name)(nil),
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
@@ -364,7 +449,7 @@ func file_tests_message_message_conformance_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_tests_message_message_conformance_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   4,
+			NumMessages:   5,
 			NumExtensions: 0,
 			NumServices:   0,
 		},