+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Enum values nested under renamed messages will now be correctly renamed.
 ## [v0.2.0] — 2020-12-28
 
 ### Breaking Changes